requirements.txt
Flask==2.2.5
flask-cors==3.0.10
psycopg2-binary
gunicorn==20.1.0
requests==2.31.0